+++ tests/SlmQueueTest/Job/JobPluginManagerTest.php
@@ -3,6 +3,7 @@
 namespace SlmQueueTest\Job;
 
 use PHPUnit_Framework_TestCase as TestCase;
+use SlmQueue\Job\JobPluginManager;
 use SlmQueueTest\Util\ServiceManagerFactory;
 use Zend\ServiceManager\ServiceManager;
 
@@ -33,5 +34,26 @@ class JobPluginManagerTest extends TestCase
         $secondInstance = $jobPluginManager->get('SlmQueueTest\Asset\SimpleJob');
 
         $this->assertNotSame($firstInstance, $secondInstance);
+    }
+
+    public function testPluginManagerSetsServiceNameAsMetadata()
+    {
+        $jobPluginManager = new JobPluginManager;
+        $jobPluginManager->setInvokableClass('SimpleJob', 'SlmQueueTest\Asset\SimpleJob');
+
+        $instance = $jobPluginManager->get('SimpleJob');
+
+        $this->assertInstanceOf('SlmQueueTest\Asset\SimpleJob', $instance);
+        $this->assertEquals('SimpleJob', $instance->getMetadata('name'));
+    }
+
+    public function testPluginManagerThrowsExceptionOnInvalidJobClasses()
+    {
+        $jobPluginManager = new JobPluginManager;
+        $jobPluginManager->setInvokableClass('InvalidJob', 'stdClass');
+
+        $this->setExpectedException('SlmQueue\Job\Exception\RuntimeException');
+
+        $instance = $jobPluginManager->get('InvalidJob');
     }
 }